Subiendo a master la ultima version que fue aprobada en el Final de Trabajo de Diploma diciembre 2014

+++ AppCursoAspNet/AppCursoAspNet/View/Choferes.aspx
@@ -1,13 +1,20 @@
 ﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site1.Master" AutoEventWireup="true" CodeBehind="Choferes.aspx.cs" Inherits="Vista.View.Choferes" %>
 
 <asp:Content ID="BodyContent" ContentPlaceHolderID="MainContent" runat="server">
-  <div>
+    <div>
         <ol class="breadcrumb">
             <li><a href="Index.aspx">Inicio</a></li>
             <li class="active">Choferes</li>
         </ol>
     </div>
     <h2>Choferes</h2>
+
+        <p>
+        <asp:HyperLink ID="HyperLink1" runat="server" NavigateUrl="~/View/NuevoChofer.aspx" class="btn"><span class="glyphicon glyphicon-plus"></span>Agregar Chofer</asp:HyperLink>
+    </p>
+        <form runat="server">
+                <asp:ScriptManager ID="ScriptManager1" AllowCustomErrorsRedirect="false" runat="server"></asp:ScriptManager>
+
 
     <asp:ListView ID="listChoferes" runat="server" GroupItemCount="3" ItemType="Modelo.Chofer" DataSourceID="ObjectChofer">
         <EmptyDataTemplate>
@@ -37,8 +44,8 @@
                         </td>
                     </tr>
                     <tr>
-                        <td>
-                            <code><%#:Item.Nombre%> <%#:Item.Apellido%></code>
+                        <td class="tdChofer">
+                            <code><%#:Item.NombreCompleto%></code>
                             <br />
                             <span>
                                 <b>DNI: </b><%#Item.Documento.ToString()%>
@@ -49,7 +56,7 @@
                             </span>
                             <br />
                             <span>
-                                <b>Fecha de Nacimiento: </b><%#:Item.FechaNacimiento%>
+                                <b>Fecha de Nacimiento: </b><%#:Item.FechaNacimientoCorta%>
                             </span>
                             <br />
                             <span>
@@ -69,12 +76,12 @@
                             </span>
                             <br />
                             
-                            <a href="/View/EliminarChofer.aspx?Documento=<%#:Item.Documento%>" role="button" >
-                                <span class="glyphicon glyphicon-delete"></span>
+                            <a href="/View/EliminarChofer.aspx?Documento=<%#:Item.Documento%>" role="button" id="btnEliminar" class="btnEliminar" onclick="return confirm('Estas seguro que deseas eliminar este registro?');">
+                                <span class="glyphicon glyphicon-delete yellow-icon"></span>
                                     <b>Eliminar<b/>
                             </a><br/>
-                            <a href="/View/EditarChofer.aspx?Documento=<%#:Item.Documento%>" role="button" >
-                                <span class="glyphicon "></span>
+                            <a href="/View/EditarChofer.aspx?Documento=<%#:Item.Documento%>" role="button" id="btnEditar" class="btnEditar">
+                                <span class="glyphicon glyphicon-edit yellow-icon"></span>
                                     <b>Editar<b/>
                             </a><br/>
                         </td>
@@ -104,10 +111,35 @@
             </table>
         </LayoutTemplate>
     </asp:ListView>
-    <asp:ObjectDataSource ID="ObjectChofer" runat="server" SelectMethod="ListarChoferes" TypeName="Controladora.ControladoraChoferes"></asp:ObjectDataSource>
-    
 
-
+    <asp:ObjectDataSource ID="ObjectChofer" runat="server" SelectMethod="ListChoferes_GetData" TypeName="Vista.View.Choferes"></asp:ObjectDataSource>
+           
+            
+            
+                <!-- Bootstrap Modal Dialog -->
+<div class="modal fade" id="myModal" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <asp:UpdatePanel ID="upModal" runat="server" ChildrenAsTriggers="false" UpdateMode="Conditional">
+            <ContentTemplate>
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+                        <h4 class="modal-title"><asp:Label ID="lblModalTitle" runat="server" Text=""></asp:Label></h4>
+                    </div>
+                    <div class="modal-body">
+                        <asp:Label ID="lblModalBody" runat="server" Text=""></asp:Label>
+                    </div>
+                    <div class="modal-footer">
+                        <button class="btn btn-amarillo" data-dismiss="modal" aria-hidden="true">Close</button>
+                    </div>
+                </div>
+            </ContentTemplate>
+        </asp:UpdatePanel>
+    </div>
+</div>
+            
+            
+             </form>
